package.json and file restructure.
