+++ blockly-rtc/public/index.html
@@ -20,6 +20,9 @@
 </head>
 
 <body>
+    <div id='blocklyDiv'>
+    </div>
+
     <script type='text/javascript' src='./main.js'></script>
 
     <xml xmlns="https://developers.google.com/blockly/xml" id="toolbox" style="display: none">
