+++ blockly-rtc/test/mocha.opts
@@ -1,0 +1,3 @@
+# mocha.opts
+--ui tdd
+-r esm
